Move functions to labeler utilities
