+++ DataScience/GpxAudioLabeler.py
@@ -9,10 +9,8 @@ from lxml import etree
 from matplotlib import pyplot as plt
 import numpy as np
 import pandas as pd
-from sklearn.cluster import KMeans
-from sklearn_extra.cluster import KMedoids
 
-import Utilities as u
+import LabelerUtilities as lu
 
 # WGS84 parameters
 R_OPLUS = 6378137  # [m]
@@ -32,7 +30,7 @@ logger = logging.getLogger("GpxAudioLabeler")
 logger.setLevel(logging.INFO)
 
 
-def parse_source_gpx_file(inp_path):
+def parse_source_gpx_file(inp_path, source):
     """Parse a GPX file having the following structure:
 
     <gpx xmlns="http://www.topografix.com/GPX/1/1" xmlns:gpxtpx="http://www.garmin.com/xmlschemas/TrackPointExtension/v1" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" creator="Suunto app" version="1.1" xsi:schemaLocation="http://www.topografix.com/GPX/1/1 http://www.topografix.com/GPX/1/1/gpx.xsd http://www.garmin.com/xmlschemas/TrackPointExtension/v1 http://www.garmin.com/xmlschemas/TrackPointExtensionv1.xsd">
@@ -63,11 +61,21 @@ def parse_source_gpx_file(inp_path):
     ----------
     inp_path : pathlib.Path()
         Path of the GPX file to parse
+    source : dict
+        The source configuration
 
     Returns
     -------
     gpx : dict
         Track, track segments, and track points
+    vld_t
+        Time from start of track [s]
+    vld_lambda
+        Geodetic longitude [rad]
+    vld_varphi
+        Geodetic latitude [rad]
+    vld_h
+        Elevation [m]
 
     See also:
     https://en.wikipedia.org/wiki/GPS_Exchange_Format
@@ -135,220 +143,34 @@ def parse_source_gpx_file(inp_path):
 
         gpx["trks"].append(trk)
 
-    return gpx
+        # Assign longitude, latitude, elevation, and time from start of
+        # track
+        # TODO: Check single track and track segment assumption
+        _t = np.array(
+            gpx["trks"][0]["trksegs"][0]["time"]
+        )  # time from start of track [s]
+        _lambda = np.array(
+            gpx["trks"][0]["trksegs"][0]["lon"]
+        )  # geodetic longitude [rad]
+        _varphi = np.array(
+            gpx["trks"][0]["trksegs"][0]["lat"]
+        )  # geodetic latitude [rad]
+        _h = np.array(gpx["trks"][0]["trksegs"][0]["ele"])  # elevation [m]
 
+        # Ignore points at which the elevation was not recorded
+        vld_idx = np.logical_and(
+            np.logical_and(source["start_t"] < _t, _t < source["stop_t"]),
+            _h != -R_OPLUS,
+        )
+        logger.info(
+            f"Found {np.sum(vld_idx)} valid values out of all {_t.shape[0]} values"
+        )
+        vld_t = _t[vld_idx]
+        vld_lambda = _lambda[vld_idx]
+        vld_varphi = _varphi[vld_idx]
+        vld_h = _h[vld_idx]
 
-def compute_source_metrics(source, gpx, hydrophone):
-    """Compute the topocentric position and velocity of the source
-    relative to the hydrophone, and corresponding heading, heading
-    first derivative, distance, and speed.
-
-    Parameters
-    ----------
-    source : dict
-        The source configuration
-    gpx : dict
-        Track, track segments, and track points
-    hydrophone : dict
-        The hydrophone configuration
-
-    Returns
-    -------
-    vld_t : numpy.ndarray
-        Time from start of track [s]
-    distance : numpy.ndarray
-        Source distance from hydrophone [m]
-    heading : numpy.ndarray
-        Source heading, zero at north, clockwise positive
-    heading_dot : numpy.ndarray
-        Source heading first derivative
-    speed : numpy.ndarray
-        Source speed [m/s]
-    r_s_h : numpy.ndarray
-        Source topocentric (east, north, zenith) position [m]
-    v_s_h : numpy.ndarray
-        Source topocentric (east, north, zenith) velocity [m/s]
-
-    See:
-    Montenbruck O., Gill E.; Satellite Orbits; Springer, Berlin (2001); pp. 37 and 188.
-    """
-    logger.info(
-        f"Computing source {source['name']} metrics for hydrophone {Path(hydrophone['name'].lower()).stem}"
-    )
-    # Assign longitude, latitude, elevation, and time from start of
-    # track
-    # TODO: Check single track and track segment assumption
-    _lambda = np.array(gpx["trks"][0]["trksegs"][0]["lon"])  # geodetic longitude [rad]
-    _varphi = np.array(gpx["trks"][0]["trksegs"][0]["lat"])  # geodetic latitude [rad]
-    _h = np.array(gpx["trks"][0]["trksegs"][0]["ele"])  # elevation [m]
-    _t = np.array(gpx["trks"][0]["trksegs"][0]["time"])  # time from start of track [s]
-
-    # Ignore points at which the elevation was not recorded
-    vld_idx = np.logical_and(
-        np.logical_and(source["start_t"] < _t, _t < source["stop_t"]), _h != -R_OPLUS
-    )
-    logger.info(f"Found {np.sum(vld_idx)} valid values out of all {_t.shape[0]} values")
-    vld_lambda = _lambda[vld_idx]
-    vld_varphi = _varphi[vld_idx]
-    vld_h = _h[vld_idx]
-    vld_t = _t[vld_idx]
-
-    # Assign longitude, latitude, and elevation of hydrophone
-    hyd_lambda = math.radians(hydrophone["lon"])
-    hyd_varphi = math.radians(hydrophone["lat"])
-    hyd_h = math.radians(hydrophone["ele"])
-
-    # Compute the orthogonal transformation matrix from geocentric to
-    # topocentric coordinates at hydrophone longitude, and latitude
-    E = u.compute_E(hyd_lambda, hyd_varphi)
-
-    # Compute the geocentric position of the hydrophone, source, and
-    # source relative to the hydrophone
-    R_h = u.compute_R(hyd_lambda, hyd_varphi, hyd_h)
-    R_s = u.compute_R(vld_lambda, vld_varphi, vld_h)
-    R_s_h = R_s - np.atleast_2d(R_h).reshape(3, 1)
-
-    # Compute the topocentric position and velocity of the source
-    # relative to the origin, and corresponding heading, heading first
-    # derivative, distance, and speed
-    # TODO: Use instantaneous orthogonal transformation matrix?
-    r_s_h = np.matmul(E, R_s_h)
-    v_s_h = np.gradient(r_s_h, vld_t, axis=1)
-    distance = np.sqrt(np.sum(np.square(r_s_h), axis=0))
-    heading = 90 - np.degrees(np.arctan2(v_s_h[1, :], v_s_h[0, :]))
-    heading_dot = np.abs(
-        pd.DataFrame(np.gradient(heading, vld_t)).ewm(span=3).mean().to_numpy()
-    ).flatten()
-    speed = np.sqrt(np.sum(np.square(v_s_h), axis=0))
-    return vld_t, distance, heading, heading_dot, speed, r_s_h, v_s_h
-
-
-def plot_source_metrics(
-    source, hydrophone, heading, heading_dot, distance, speed, r_s_h
-):
-    """Plot source track, and histograms of source distance, heading,
-    heading first derivative, and speed.
-
-    Parameters
-    ----------
-    source : dict
-        The source configuration
-    hydrophone : dict
-        The hydrophone configuration
-    heading : numpy.ndarray
-        Source heading, zero at north, clockwise positive
-    heading_dot : numpy.ndarray
-        Source heading first derivative
-    distance : numpy.ndarray
-        Source distance from hydrophone [m]
-    speed : numpy.ndarray
-        Source speed [m/s]
-    r_s_h : numpy.ndarray
-        Source topocentric (east, north, zenith) position [m]
-
-    Returns
-    -------
-    None
-    """
-    logger.info(
-        f"Plotting source {source['name']} metrics for hydrophone {Path(hydrophone['name'].lower()).stem}"
-    )
-    fig, axs = plt.subplots()
-    axs.plot(r_s_h[0, :], r_s_h[1, :])
-    axs.axhline(color="gray", linestyle="dotted")
-    axs.axvline(color="gray", linestyle="dotted")
-    axs.set_title("Track")
-    axs.set_xlabel("east [m]")
-    axs.set_ylabel("north [m]")
-    plt.show()
-
-    fig, axs = plt.subplots()
-    axs.hist(distance, bins=100)
-    axs.set_title("Distance")
-    axs.set_xlabel("distance [m]")
-    axs.set_ylabel("counts")
-    plt.show()
-
-    fig, axs = plt.subplots()
-    axs.hist(90 - heading, bins=180)
-    axs.set_title("Headings")
-    axs.set_xlabel("heading [deg]")
-    axs.set_ylabel("counts")
-    plt.show()
-
-    fig, axs = plt.subplots()
-    axs.hist(np.abs(heading_dot), bins=100)
-    axs.set_title("Heading First Derivative")
-    axs.set_xlabel("heading first derivative [deg/s]")
-    axs.set_ylabel("counts")
-    plt.show()
-
-    fig, axs = plt.subplots()
-    axs.hist(speed, bins=100)
-    axs.set_title("Speed")
-    axs.set_xlabel("speed [m/s]")
-    axs.set_ylabel("counts")
-    plt.show()
-
-
-def cluster_source_metrics(
-    distance,
-    distance_n_clusters,
-    heading,
-    heading_n_clusters,
-    heading_dot,
-    heading_dot_n_clusters,
-    speed,
-    speed_n_clusters,
-):
-    """Compute clusters of distance, heading, heading first derivative,
-    and speed.
-
-    Parameters
-    ----------
-    distance : numpy.ndarray
-        Source distance from the hydrophone [m]
-    distance_n_clusters : int
-        Number of distance clusters
-    heading : numpy.ndarray
-        Source heading, zero at north, clockwise positive [deg]
-    heading_n_clusters : int
-        Number of heading clusters
-    heading_dot : numpy.ndarray
-        Source heading first derivative, zero at north, clockwise
-        positive [deg/s]
-    heading_dot_n_clusters : int
-        Number of heading dot clusters
-    speed : numpy.ndarray
-        Source speed [m/s]
-    speed_n_clusters : int
-        Number of speed clusters
-
-    Returns
-    -------
-    distance_clusters : sklearn.cluster.KMeans
-        Fitted estimator for distance
-    heading_clusters : sklearn.cluster.KMeans
-        Fitted estimator for heading
-    heading_dot_clusters : sklearn_extra.cluster.KMedoids
-        Fitted estimator for heading first derivative
-    speed_clusters : sklearn.cluster.KMeans
-        Fitted estimator for speed
-    """
-    logger.info(f"Computing clusters of heading, distance, and speed")
-    distance_clusters = KMeans(n_clusters=distance_n_clusters, random_state=0).fit(
-        distance.reshape(-1, 1)
-    )
-    heading_clusters = KMeans(n_clusters=heading_n_clusters, random_state=0).fit(
-        90 - heading.reshape(-1, 1)
-    )
-    heading_dot_clusters = KMedoids(
-        n_clusters=heading_dot_n_clusters, random_state=0
-    ).fit(heading_dot.reshape(-1, 1))
-    speed_clusters = KMeans(n_clusters=speed_n_clusters, random_state=0).fit(
-        speed.reshape(-1, 1)
-    )
-    return distance_clusters, heading_clusters, heading_dot_clusters, speed_clusters
+    return gpx, vld_t, vld_lambda, vld_varphi, vld_h
 
 
 def slice_source_audio_by_cluster(
@@ -504,7 +326,7 @@ def slice_source_audio_by_cluster(
                             wav_filename = (
                                 "{:s}-{:d}-{:d}{:+.1f}{:+.1f}{:+.1f}{:+.1f}{:+.1f}.wav"
                             )
-                            u.export_audio_clip(
+                            lu.export_audio_clip(
                                 audio,
                                 start_t,
                                 stop_t,
@@ -683,7 +505,7 @@ def slice_source_audio_by_condition(
             stop_t = min(hyd_min_stop_t, dub_stop_t)
             n_clips += 1
             wav_filename = "{:s}-{:d}-{:d}-{:+.1f}to{:+.1f}-{:+.1f}to{:+.1f}-and-{:+.1f}to{:+.1f}-{:+.1f}to{:+.1f}-{:+.1f}to{:+.1f}.wav"
-            u.export_audio_clip(
+            lu.export_audio_clip(
                 audio,
                 start_t,
                 stop_t,
@@ -867,7 +689,7 @@ def main():
 
     # Load file describing the collection
     collection_path = Path(args.data_home) / args.collection_filename
-    collection = u.load_json_file(collection_path)
+    collection = lu.load_json_file(collection_path)
 
     # Identify the interval during which any source emitted
     src_min_start_t = min([h["start_t"] for h in collection["sources"]]) * 1000  # [ms]
@@ -882,40 +704,48 @@ def main():
     )  # [ms]
 
     # Load file describing sampling cases
-    sampling = u.load_json_file(args.sampling_filepath)
+    sampling = lu.load_json_file(args.sampling_filepath)
 
     # Consider each course
     for source in collection["sources"]:
+        if source["type"] != "file":
+            raise Exception("Unexpected source type")
         gpx_path = Path(args.data_home) / source["name"]
-        gpx = parse_source_gpx_file(gpx_path)
+        gpx, vld_lambda, vld_varphi, vld_h, vld_t = parse_source_gpx_file(
+            gpx_path, source
+        )
 
         # Consider each hydrophone
         for hydrophone in collection["hydrophones"]:
+            if hydrophone["type"] != "file":
+                raise Exception("Unexpected hydrophone type")
             wav_path = Path(args.data_home) / hydrophone["name"]
-            audio = u.get_wav_file(wav_path)
+            audio = lu.get_wav_file(wav_path)
 
             # Export audio with no source present, if it exists
             if src_max_stop_t < hyd_min_stop_t:
-                u.export_audio_clip(
+                lu.export_audio_clip(
                     audio,
                     src_max_stop_t,
                     hyd_min_stop_t,
-                    Path(args.clip_home) / "no-boat"
+                    Path(args.clip_home)
+                    / "no-boat"
                     / f"{Path(hydrophone['name'].lower()).stem}-{src_max_stop_t}-{hyd_min_stop_t}-no-source.wav",
                 )
 
             # Compute and plot source metrics for the current hydrophone
             (
-                vld_t,
                 distance,
                 heading,
                 heading_dot,
                 speed,
                 r_s_h,
                 v_s_h,
-            ) = compute_source_metrics(source, gpx, hydrophone)
+            ) = lu.compute_source_metrics(
+                source, vld_t, vld_lambda, vld_varphi, vld_h, hydrophone
+            )
             if args.do_plot_metrics:
-                plot_source_metrics(
+                lu.plot_source_metrics(
                     source, hydrophone, heading, heading_dot, distance, speed, r_s_h
                 )
 
@@ -931,7 +761,7 @@ def main():
                         heading_clusters,
                         heading_dot_clusters,
                         speed_clusters,
-                    ) = cluster_source_metrics(
+                    ) = lu.cluster_source_metrics(
                         distance,
                         method["distance_n_clusters"],
                         heading,
